Get font from UEFI #25
