+++ src/bootx64/src/uefi/protocols/human_interface_infrastructure.rs
@@ -3,3 +3,13 @@
 // 34 HII (Human Interface Infrastructure) Protocols
 
 pub mod font;
+pub mod font_ex;
+pub mod string;
+pub mod database;
+
+// EFI_HII_FONT_STYLE
+pub type FontStyle = u32;
+
+// EFI_STRING_ID
+pub type StringId = u16;
+
